Make int32 and int64 available

+++ params.f90
@@ -1,6 +1,6 @@
 module params
 
-  use iso_fortran_env, only : np => real64
+  use iso_fortran_env, only : np => real64, int32, int64
 
   implicit none
 
@@ -8,34 +8,27 @@ module params
   real(np), parameter :: pi = acos(-1.0_np)
   real(np), parameter :: dtor = pi/180.0_np
 
-  ! coordinate system type
-  ! now set using the make file
-  ! (1) = Cartesian (x,y,z)
-  ! (2) = Spherical (r,theta,phi) (theta=polar angle, phi=azimuthal angle)
-  ! (3) = Cylindrical (rho,phi,z)
-  ! integer, parameter :: coord_type = 2
-
   ! data file containing the magnetic field data
   character(100), parameter :: defaultfilename = 'data/magfield.dat'
   character(100), parameter :: outputdir = 'output'
   character(100) :: filein, fileout
 
   ! Number of threads to be used by OpenMP
-  integer, parameter :: nproc = 4
+  integer(int32), parameter :: nproc = 4
 
   ! nullfinder parameters
   real(np), parameter :: zero = 1e-10_np ! what the code treats as zero
-  integer, parameter :: sig_figs = 6 ! the number of significant figures of accuracy required for the null
+  integer(int32), parameter :: sig_figs = 6 ! the number of significant figures of accuracy required for the null
 
   ! spinefinder parameters
   real(np), parameter :: rsphere = 1e-4_np ! radius of sphere which the dot products are calculated on
-  integer, parameter :: nphi = 90 ! number of points on sphere to start
-  integer, parameter :: ntheta = nphi/2
+  integer(int32), parameter :: nphi = 90 ! number of points on sphere to start
+  integer(int32), parameter :: ntheta = nphi/2
 
   ! ssfind parameters
-  integer, parameter :: nstart = 100 ! number of startpoints in ring
-  integer, parameter :: ringsmax = 10000 ! maximum # of rings
-  integer, parameter :: pointsmax = 100000 ! maximum number of points in ring
+  integer(int32), parameter :: nstart = 100 ! number of startpoints in ring
+  integer(int32), parameter :: ringsmax = 10000 ! maximum # of rings
+  integer(int32), parameter :: pointsmax = 100000 ! maximum number of points in ring
   real(np), parameter :: tol = 1e-5_np ! tolerance of rkf45 scheme
   real(np), parameter :: stepmin = 3e-4_np ! minimum step length
 
@@ -44,7 +37,7 @@ module params
     subroutine filenames
 
       character(100) :: arg, outname
-      integer :: iarg, ifname, ic, oc
+      integer(int32) :: iarg, ifname, ic, oc
 
       ic = 0
       oc = 0
